Former-commit-id: 3eb4f7ac8cfc78a467d06046777d776577d0ab1c

+++ src/BDSOutput.cc
@@ -614,7 +614,7 @@ void BDSOutput::FillSamplerHits(const BDSHitsCollectionSampler* hits,
   // TODO - cylinder output will have all the same z and S, which is wrong!
   if (!(hits->entries() > 0))
     {return;}
-  for (int i = 0; i < hits->entries(); i++)
+  for (int i = 0; i < (int)hits->entries(); i++)
     {
       const BDSHitSampler* hit = (*hits)[i];
       G4int samplerID = hit->samplerID;
